requirements.txt
fpdf2==2.8.5
natsort==8.4.0
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
